% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{deviance.fastglm}
\alias{deviance.fastglm}
\title{deviance method for fastglm fitted objects}
\usage{
\method{deviance}{fastglm}(object, ...)
}
\arguments{
\item{object}{fastglm fitted object}

\item{...}{not used}
}
\value{
The value of the deviance extracted from the object
}
\description{
deviance method for fastglm fitted objects
}
