% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory_methods.R
\docType{methods}
\name{\%*\%,big.matrix,vector-method}
\alias{\%*\%,big.matrix,vector-method}
\alias{\%*\%,vector,big.matrix-method}
\title{big.matrix prod}
\usage{
\S4method{\%*\%}{big.matrix,vector}(x, y)

\S4method{\%*\%}{vector,big.matrix}(x, y)
}
\arguments{
\item{x}{big.matrix}

\item{y}{numeric vector}
}
\description{
big.matrix prod

big.matrix prod
}
