% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{family.fastglm}
\alias{family.fastglm}
\title{family method for fastglm fitted objects}
\usage{
\method{family}{fastglm}(object, ...)
}
\arguments{
\item{object}{fastglm fitted object}

\item{...}{not used}
}
\value{
returns the family of the fitted object
}
\description{
family method for fastglm fitted objects
}
