% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{logLik.fastglm}
\alias{logLik.fastglm}
\title{logLik method for fastglm fitted objects}
\usage{
\method{logLik}{fastglm}(object, ...)

\method{logLik}{fastglm}(object, ...)
}
\arguments{
\item{object}{fastglm fitted object}

\item{...}{not used}

\item{object}{fastglm fitted object}

\item{...}{not used}
}
\value{
Returns an object of class \code{logLik}

Returns an object of class \code{logLik}
}
\description{
logLik method for fastglm fitted objects

logLik method for fastglm fitted objects
}
