% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{print.fastglm}
\alias{print.fastglm}
\title{print method for fastglm objects}
\usage{
\method{print}{fastglm}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{not used}
}
\description{
print method for fastglm objects
}
