% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{residuals.fastglm}
\alias{residuals.fastglm}
\title{residuals method for fastglm fitted objects}
\usage{
\method{residuals}{fastglm}(object, type = c("deviance", "pearson",
  "working", "response", "partial"), ...)
}
\arguments{
\item{object}{fastglm fitted object}

\item{type}{type of residual to be returned}

\item{...}{not used}
}
\value{
a vector of residuals
}
\description{
residuals method for fastglm fitted objects
}
