% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglm_nb.R, R/glm_methods.R
\name{vcov.fastglm_nb}
\alias{vcov.fastglm_nb}
\alias{vcov.fastglm}
\title{vcov method for fastglm_nb fitted objects}
\usage{
\method{vcov}{fastglm_nb}(object, ...)

\method{vcov}{fastglm}(object, ...)
}
\arguments{
\item{object}{fastglm fitted object}

\item{...}{not used}
}
\value{
the covariance matrix

the covariance matrix
}
\description{
vcov method for fastglm_nb fitted objects

vcov method for fastglm fitted objects
}
