% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglm-methods.R
\name{vcovCL.fastglm}
\alias{vcovCL.fastglm}
\alias{vcovCL.fastglmFit}
\title{Cluster-robust variance estimator for `fastglm` objects}
\usage{
vcovCL.fastglm(object, cluster, type = c("HC1", "HC0"), ...)

vcovCL.fastglmFit(object, cluster, type = c("HC1", "HC0"), ...)
}
\arguments{
\item{object}{a fitted object of class `"fastglm"` or `"fastglmFit"`.}

\item{cluster}{a vector identifying the cluster for each observation. Length
must equal the number of rows of the design matrix.}

\item{type}{one of `"HC0"`, `"HC1"`. `"HC1"` (the default) applies the small-sample
degrees-of-freedom adjustment `(G / (G - 1)) * ((n - 1) / (n - p))`, where
`G` is the number of clusters.}

\item{...}{not used.}
}
\value{
A `p x p` cluster-robust variance-covariance matrix.
}
\description{
Methods for `sandwich::vcovCL()` on objects of class `"fastglm"` and
`"fastglmFit"`. Load `sandwich` (`library(sandwich)`) before calling
`vcovCL(fit, cluster)`; otherwise no `vcovCL` generic is in scope.
}
\details{
The Liang-Zeger (1986) cluster-robust sandwich:
`bread %*% (sum_g s_g s_g') %*% bread`, where `s_g = sum_{i in g} w_i r_i x_i`
is the score contribution from cluster `g`.
}
