% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_fit_glm.R
\name{fastglmPure}
\alias{fastglmPure}
\title{fast and memory efficient linear model fitting}
\usage{
fastglmPure(X, y, weights = rep(1, NROW(y)), offset = rep(0, NROW(y)),
  family = gaussian(), type = 0L, tol = 1e-07, maxit = 100L)
}
\arguments{
\item{X}{input model matrix. Must be a matrix object}

\item{y}{numeric response vector of length nobs.}

\item{weights}{an optional vector of 'prior weights' to be used in the fitting process. Should be a numeric vector.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting. 
This should be a numeric vector of length equal to the number of cases}

\item{type}{an integer scalar with value 0 for the column-pivoted QR decomposition, 1 for the unpivoted QR decomposition,
2 for the LLT Cholesky, or 3 for the LDLT Cholesky}

\item{tol}{threshold tolerance for convergence. Should be a positive real number}

\item{maxit}{maximum number of IRLS iterations. Should be an integer}
}
\value{
A list with the elements
\item{coefficients}{a vector of coefficients}
\item{se}{a vector of the standard errors of the coefficient estimates}
\item{rank}{a scalar denoting the computed rank of the model matrix}
\item{df.residual}{a scalar denoting the degrees of freedom in the model}
\item{residuals}{the vector of residuals}
\item{s}{a numeric scalar - the root mean square for residuals}
\item{fitted.values}{the vector of fitted values}
}
\description{
fast and memory efficient linear model fitting
}
