% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{vcov.fastglm}
\alias{vcov.fastglm}
\title{vcov method for fastglm fitted objects}
\usage{
\method{vcov}{fastglm}(object, ...)
}
\arguments{
\item{object}{fastglm fitted object}

\item{...}{not used}
}
\value{
the covariance matrix
}
\description{
vcov method for fastglm fitted objects
}
